:truck: Move command runners package into service package
