--- src/core/commands/defaults/runners/__init__.py
@@ -1,4 +1,0 @@
-from .server import ServerRunner
-from .test import TestRunner
-
-__all__ = ["ServerRunner", "TestRunner"]
